requirements.txt
Flask==2.3.2
pandas==2.2.3
numpy==1.26.4
openpyxl==3.1.2
matplotlib==3.8.1
Pillow==11.3.0
setuptools>=65.0.0
